Move `ProviderFactory` and dependent service `JavaInstallationRegistry` down
To the build scope so `ProviderFactory.gradleProperty` has access to the
`BuildLayout` and `IGradlePropertiesLoader`.
